Refactor CLIFF to use a plugin system for Entity Extraction
This refactor makes the entity extraction for CLIFF pluggable. This allows
the use of non-GNU extraction methods such as the MITIE library. It also
should allow the use of multiple targeted extraction methods. For example
if plugin one can support types X and Y, and plugin two type Z, this
should be able to merge the two entity sets.